Fix Python 3 doctests Unicode escape string `DeprecationWarning` warning
When testing `http.rst`, `test_docs.py` outputs a `DeprecationWarning` because
the example text is parsed as a Unicode string by Python 3 and `\$` (meant as a
shell escape) is interpreted as an invalid Unicode escape sequence. This change
fixes the issue by instructing Python to parse the example text as a raw
string (thus ignoring and correctly passing `\$` through to the shell).

See: https://docs.python.org/3/reference/lexical_analysis.html#string-and-bytes-literals